+++ DaleGhent.NINA.AstroPhysics/CHANGELOG.md
@@ -1,5 +1,15 @@
 ﻿# Utilities for Astro-Physics mounts
 
+## 1.3.0.0 - 2021-9-1
+* Added instruction: Start APCC
+    - Starts APCC and connects NINA to the Astro-Physics ASCOM driver. Requires the following:
+        * **Auto-Connect** is selected in APCC's Setup tab for both the **Mount** and **AP V2 Driver**
+        * The **Astro-Physics GTO V2 Mount** driver is already selected and saved in the active NINA profile
+
+## 1.2.0.0 - unreleased
+* Updated plugin description to markdown format
+* Minimum supported NINA version is now 1.11 build 125
+
 ## 1.1.0.0 - 2021-8-1
 * Minimum supported NINA version is now 1.11 build 116